Add script to genrate dataset for presentation
Large datasets may be too overwhelming and time consuming on some systems.
This shell script takes the 2nd Linked Open Data-enabled Recommender
Systems Challenge dataset converts userID and itemID to integers and gives
random ratings between 1 to 5 for the training entries.
The genrated dataset is in no way ment for evaluation of the RS.