Add page with example RDF query to DBpedia
An RDF query language is a computer language, specifically a query language for databases, able to retrieve and manipulate data stored in Resource Description Framework format.
SPARQL is emerging as the de facto RDF query language, and is a W3C recommendation.

See:
  https://en.wikipedia.org/wiki/RDF_query_language
  https://en.wikipedia.org/wiki/SPARQL

This commit adds a PHP page, 'DBpediaPull.php' which makes use of the ARC2 pligin for SPARQL support. Queries are made to DBpedia's Remote Store Endpoint. All PHP related dependencies are managed via Composer.

See:
  https://github.com/semsol/arc2/wiki/Remote-Stores-and-Endpoints
  https://getcomposer.org/doc/00-intro.md

This commit also adds favicons for the project.